review: fixes for a few command responses.
also one warning addressed in waiter.cpp

+++ src/win32/waiter.cpp
@@ -108,7 +108,7 @@ bool WinWaiter::addhandle(HANDLE handle, int flag)
     assert(handles.size() == flags.size() && handles.size() >= index);
 
 #ifdef DEBUG
-    for (unsigned i = index; i--; )
+    for (size_t i = index; i--; )
     {
         // double check we only add one of each handle
         assert(handles[i] != handle);
